feat: refactor item loading logic and create Item class for better song management
